# api/openapi.yaml
openapi: 3.1.0
info:
  title: tyrs-hand Management API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
...
paths:
...
  /development-environments:
    get:
      operationId: listDevelopmentEnvironments
      tags: [Development Environments]
      responses:
        "200":
          description: 长期开发环境
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DevelopmentEnvironmentList" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      operationId: createDevelopmentEnvironment
      tags: [Development Environments]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ownerDiscordUserId]
              properties:
                ownerDiscordUserId: { type: string, minLength: 1 }
      responses:
        "201":
          description: 逻辑环境已创建并绑定 Worker
          content: { application/json: { schema: { $ref: "#/components/schemas/IDResource" } } }
        default: { $ref: "#/components/responses/Problem" }
...
  /workers:
    get:
      operationId: listHostWorkers
      tags: [Workers]
      responses:
        "200":
          description: 宿主 Worker 列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties: { items: { type: array, items: { $ref: "#/components/schemas/ExecutionNode" } } }
    post:
      operationId: createHostWorker
      tags: [Workers]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ExecutionNodeInput" } } }
      responses:
        "201":
          description: Worker 与一次性注册 Token
          content: { application/json: { schema: { $ref: "#/components/schemas/ExecutionNodeCreated" } } }
...
  /workers/{id}:
    delete:
      operationId: deleteHostWorker
      tags: [Workers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/CSRFToken" }
      responses: { "204": { description: 已删除 }, default: { $ref: "#/components/responses/Problem" } }
...
  /worker/v2/enroll:
    post:
      operationId: enrollWorker
      tags: [Worker Protocol]
      servers: [{ url: / }]
      security: []
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/WorkerEnrollRequest" } } } }
      responses: { "200": { description: 长期节点凭据, content: { application/json: { schema: { $ref: "#/components/schemas/WorkerEnrollResponse" } } } }, default: { $ref: "#/components/responses/Problem" } }
  /worker/v2/sync:
    post:
      operationId: syncWorker
      tags: [Worker Protocol]
      servers: [{ url: / }]
      security: [{ nodeBearer: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/WorkerRequestEnvelope" } } } }
      responses:
        "200": { description: 同步结果, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        "204": { description: 同步完成，无响应体 }
        default: { $ref: "#/components/responses/Problem" }
  /worker/v2/rpc:
    post:
      operationId: callWorkerRPC
      tags: [Worker Protocol]
      servers: [{ url: / }]
      security: [{ nodeBearer: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/WorkerRequestEnvelope" } } } }
      responses:
        "200": { description: RPC 结果, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        "204": { description: RPC 完成，无响应体 }
        default: { $ref: "#/components/responses/Problem" }
  /worker/v2/blobs/{id}:
    get:
      operationId: downloadWorkerBlob
      tags: [Worker Protocol]
      servers: [{ url: / }]
      security: [{ nodeBearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: runId, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: X-Run-Lease-Token, in: header, required: true, schema: { type: string } }
        - { name: X-Run-Lease-Epoch, in: header, required: true, schema: { type: integer, format: int64 } }
      responses: { "200": { description: 二进制附件, content: { application/octet-stream: { schema: { type: string, format: binary } } } }, default: { $ref: "#/components/responses/Problem" } }
    post:
      operationId: uploadWorkerBlob
      tags: [Worker Protocol]
      servers: [{ url: / }]
      security: [{ nodeBearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: ordinal, in: query, required: true, schema: { type: integer, minimum: 0 } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [metadata, file]
              properties:
                metadata: { type: string }
                file: { type: string, format: binary }
      responses: { "200": { description: Blob 上传结果, content: { application/json: { schema: { type: object, additionalProperties: true } } } }, default: { $ref: "#/components/responses/Problem" } }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tyrs_hand_session
    nodeBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-node-token
    clientBearer:
      type: http
      scheme: bearer
      bearerFormat: tyrs-hand-device-token
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
...
  schemas:
...
    ExecutionNode:
      type: object
      required: [id, name, roles, enabled, maxConcurrentJobs, protocolVersion, status, metadata]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        roles: { type: array, items: { type: string, enum: [github, discord] } }
        enabled: { type: boolean }
        maxConcurrentJobs: { type: integer, minimum: 1 }
        protocolVersion: { type: integer }
        workerVersion: { type: string }
        status: { type: string, enum: [pending, online, offline, disabled, incompatible, error] }
        heartbeatAt: { type: string, format: date-time }
        lastError: { type: string }
        metadata: { type: object, additionalProperties: true }
    ExecutionNodeInput:
      type: object
      required: [name, roles, maxConcurrentJobs]
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        roles: { type: array, minItems: 1, items: { type: string, enum: [github, discord] } }
        maxConcurrentJobs: { type: integer, minimum: 1 }
    EnrollmentToken:
      type: object
      required: [enrollmentToken, expiresIn]
      properties:
        enrollmentToken: { type: string, writeOnly: true }
        expiresIn: { type: integer, enum: [900] }
    ExecutionNodeCreated:
      allOf:
        - { $ref: "#/components/schemas/EnrollmentToken" }
        - type: object
          required: [node]
          properties: { node: { $ref: "#/components/schemas/ExecutionNode" } }
...
    WorkerEnrollRequest:
      type: object
      required: [token]
      properties: { token: { type: string, writeOnly: true } }
    WorkerEnrollResponse:
      type: object
      required: [nodeId, credential, protocolVersion]
      properties:
        nodeId: { type: string, format: uuid }
        credential: { type: string, writeOnly: true }
        protocolVersion: { type: integer }
    WorkerRequestEnvelope:
      type: object
      required: [requestId, sequence, operation]
      properties:
        requestId: { type: string, format: uuid }
        sequence: { type: integer, format: int64, minimum: 1 }
        operation:
          type: string
          description: 稳定操作名；sync 仅接受 worker.heartbeat 和 worker.claim。
        parameters:
          type: object
          additionalProperties: { type: string }
        payload:
          description: 操作对应的 JSON 请求体。
...
    ProblemDetails:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        requestId: { type: string }
...
    DevelopmentForumCollaborator:
      type: object
      required: [forumId, memberId, accessLevel, administratorBypass]
      properties:
        forumId: { type: string, format: uuid }
        memberId: { type: string }
        accessLevel: { type: string, enum: [readonly, operator] }
        administratorBypass: { type: boolean }
    DevelopmentForum:
      type: object
      required: [id, name, discordId, bindingStatus, collaborators]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        discordId: { type: string }
        bindingStatus: { type: string, enum: [active, inactive] }
        collaborators:
          type: array
          items: { $ref: "#/components/schemas/DevelopmentForumCollaborator" }
    DevelopmentProject:
      type: object
      required: [id, name, relativePath, projectKind, availabilityStatus, dirty, lastSeenAt, forums]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        relativePath: { type: string }
        desiredRelativePath: { type: string }
        projectKind: { type: string, enum: [directory, git] }
        availabilityStatus: { type: string, enum: [available, missing] }
        branch: { type: string }
        headSha: { type: string }
        dirty: { type: boolean }
        remoteUrl: { type: string }
        lastSeenAt: { type: string, format: date-time }
        scanError: { type: string }
        forums:
          type: array
          items: { $ref: "#/components/schemas/DevelopmentForum" }
    DevelopmentEnvironment:
      type: object
      required: [id, ownerDiscordUserId, ownerName, status, imageRef, codexUserOverride, lastUsedAt, sshConfigRevision, sshAppliedRevision, daemonStatus, appServerStatus, sshStatus, hubStatus, projects]
      properties:
        id: { type: string, format: uuid }
        ownerDiscordUserId: { type: string }
        ownerName: { type: string }
        status: { type: string }
        imageRef: { type: string }
        imageId: { type: string }
        runtimeUser: { type: string }
        codexVersion: { type: string }
        codexUserOverride: { type: boolean }
        lastUsedAt: { type: string, format: date-time }
        error: { type: string }
        executionNodeId: { type: string, format: uuid }
        sshPublicKey: { type: string }
        sshFingerprint: { type: string, readOnly: true }
        sshPort: { type: integer, minimum: 1, maximum: 65535 }
        sshDiscordUserId: { type: string }
        sshDisplayName: { type: string, readOnly: true }
        sshConfigRevision: { type: integer, format: int64, readOnly: true }
        sshAppliedRevision: { type: integer, format: int64, readOnly: true }
        daemonStatus: { type: string, enum: [pending, starting, running, error] }
        daemonError: { type: string }
        appServerStatus: { type: string, enum: [pending, starting, running, error] }
        sshStatus: { type: string, enum: [disabled, pending, starting, running, error] }
        hubStatus: { type: string, enum: [pending, starting, running, error] }
        projectsScannedAt: { type: string, format: date-time }
        projectScanError: { type: string }
        projects:
          type: array
          items: { $ref: "#/components/schemas/DevelopmentProject" }
    DevelopmentEnvironmentList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/DevelopmentEnvironment" }
...
    IDResource:
      type: object
      required: [id]
      properties:
        id: { type: string, format: uuid }
...
security:
  - sessionCookie: []